requirements.txt
selenium==4.23.1
pytest==8.3.2
webdriver-manager==4.0.2
python-dotenv==1.0.1
pytest-xdist==3.6.1